requirements.txt
fire==0.3.1
inotify==0.2.10
nose==1.3.7
PyYAML==5.3.1
six==1.14.0
termcolor==1.1.0
